requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
awsebcli==3.19.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
botocore==1.20.40
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cssutils==2.2.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
django-filebrowser-no-grappelli==3.8.0
django-inlinecss==0.3.0
Django-tinymce-filebrowser==0.2.1
django-tinymce4-lite==1.8.0
entrypoints==0.3
future==0.16.0
idna==2.10
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsmin==2.2.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.5.9
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.0
pynliner==0.8.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==22.0.3
reportlab==3.5.66
requests==2.25.1
semantic-version==2.8.5
Send2Trash==1.5.0
six==1.14.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.1.9
webencodings==0.5.1
